Run script synchronize in zygisk case

+++ native/jni/am_proc_start.cpp
@@ -109,7 +109,7 @@ int run_script(const char *arg1, const char *arg2, const char *arg3, const char 
     }
 }
 
-void run_scripts(int pid, int uid, const char *process, int user) {
+void run_scripts(bool zygisk, int pid, int uid, const char *process, int _user) {
     do {
         struct stat ppid_st, pid_st;
         char pid_str[10];
@@ -117,13 +117,13 @@ void run_scripts(int pid, int uid, const char *process, int user) {
         char user_str[10];
         snprintf(pid_str, 10, "%d", pid);
         snprintf(uid_str, 10, "%d", uid);
-        snprintf(user_str, 10, "%d", uid);
+        snprintf(user_str, 10, "%d", _user);
         int i=0;
         vector<string> module_run;
         vector<string> module_run_st2;
 
         // run script before enter the mount namespace of target app process
-        kill(pid, SIGSTOP);
+        if (!zygisk) kill(pid, SIGSTOP);
         for (auto i = 0; i < module_list.size(); i++){
             string script = "/data/adb/modules/"s + module_list[i] + "/dynmount.sh"s;
             if (access(script.data(), F_OK) != 0) continue;
@@ -132,6 +132,7 @@ void run_scripts(int pid, int uid, const char *process, int user) {
                 LOGI("run %s#prepareEnterMntNs [%s] pid=[%d] exited with code %d", module_list[i].data(), process, pid, ret/256);
             if (ret == 0) module_run.emplace_back(module_list[i]);
         }
+        if (zygisk) goto enter_mnt_ns;
         kill(pid, SIGCONT);
 
         // if there is no script we want to run in app mount namespace
@@ -155,6 +156,7 @@ void run_scripts(int pid, int uid, const char *process, int user) {
         // run script after enter the mount namespace of target app process
         // magisk module can modify mount namespce by doing mount/unmount
         kill(pid, SIGSTOP);
+        enter_mnt_ns:
         if (!switch_mnt_ns(pid)){
             for (auto i = 0; i < module_run.size(); i++){
                 string script = "/data/adb/modules/"s + module_run[i] + "/dynmount.sh"s;
@@ -168,7 +170,8 @@ void run_scripts(int pid, int uid, const char *process, int user) {
                 LOGI("no module to run OnSetUID [%s] pid=[%d]", process, pid);
                 goto unblock_process;
             }
-            kill(pid, SIGCONT);
+            if (!zygisk) kill(pid, SIGCONT);
+            if (zygisk && fork_dont_care() > 0) return;
             string path = "/proc/"s + pid_str;
             int count = 0;
             do {
@@ -193,15 +196,15 @@ void run_scripts(int pid, int uid, const char *process, int user) {
             }
         }
         unblock_process:
-        kill(pid, SIGCONT);
+        if (!zygisk) kill(pid, SIGCONT);
         return;
     } while (false);
 }
 
 
-void run_daemon(int pid, int uid, const char *process, int user){
+void run_daemon(int pid, int uid, const char *process, int _user){
     if (fork_dont_care()==0){
-        run_scripts(pid,uid,process,user);
+        run_scripts(false,pid,uid,process,_user);
         _exit(0);
     }
 }
